+++ src/main/java/com/library/entities/LigneApprovisionnement.java
@@ -25,7 +25,7 @@ public class LigneApprovisionnement implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private int numero;
+    private long numero;
     private int quantite;
     private double prix;
     private double prixAppro;
@@ -48,7 +48,7 @@ public class LigneApprovisionnement implements Serializable {
         super();
     }
 
-    public LigneApprovisionnement(Long id, int numero, int quantite, double prix, double prixAppro, Approvisionnement approvisionnement, Produit produit) {
+    public LigneApprovisionnement(Long id, long numero, int quantite, double prix, double prixAppro, Approvisionnement approvisionnement, Produit produit) {
         this.id = id;
         this.numero = numero;
         this.quantite = quantite;
@@ -65,11 +65,11 @@ public class LigneApprovisionnement implements Serializable {
         this.id = id;
     }
 
-    public Integer getNumero() {
+    public long getNumero() {
         return numero;
     }
 
-    public void setNumero(int numero) {
+    public void setNumero(long numero) {
         this.numero = numero;
     }
