+++ src/main/java/com/library/services/LigneApprovisionnementService.java
@@ -16,8 +16,8 @@ public interface LigneApprovisionnementService {
     LigneApprovisionnement updateLigneApprovisionnement(Long lApproId, LigneApprovisionnement ligneApprovisionnement);
     ResponseEntity<Object> deleteLigneApprovisionnement(Long lApproid);
 
-    void deleteLApproByNumero(int numero);
-    List<LigneApprovisionnement> findAllLApproByNumero(int numero);
+    void deleteLApproByNumero(long numero);
+    List<LigneApprovisionnement> findAllLApproByNumero(long numero);
 
     List<LigneApprovisionnement> findListLigneApprovisionnementByProduitId(Long prodId);
     List<LigneApprovisionnement> findListLigneApprovisionnementByApprovisionnementId(Long approId);
